backend/app/agent/agents/insurance_pipeline/requirements.txt
openai>=1.0.0
pydantic>=2.0.0
python-docx>=1.1.0
PyMuPDF>=1.24.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
